+++ src/test/java/com/asgeirnilsen/blog/CircularBufferTest.java
@@ -8,7 +8,7 @@ import static org.junit.Assume.assumeThat;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 import org.junit.Test;
 
@@ -17,14 +17,14 @@ public class CircularBufferTest {
     @Test
     public void emptyShouldReturnNull() {
         CircularBuffer<Integer> buf = new CircularBuffer<Integer>(10);
-        AtomicInteger idx = buf.index();
+        AtomicLong idx = buf.index();
         assertThat(buf.take(idx), nullValue());
     }
 
     @Test
     public void emptyShouldDrainEmptyList() throws Exception {
         CircularBuffer<Integer> buf = new CircularBuffer<Integer>(10);
-        AtomicInteger idx = buf.index();
+        AtomicLong idx = buf.index();
         List<Integer> result = buf.drain(idx);
         assertThat(result.size(), is(0));
     }
@@ -32,7 +32,7 @@ public class CircularBufferTest {
     @Test
     public void twoDrainsYieldPartialResults() throws Exception {
         CircularBuffer<Integer> buf = new CircularBuffer<Integer>(10);
-        AtomicInteger idx = buf.index();
+        AtomicLong idx = buf.index();
         for (int i = 0; i < 10; i++)
             buf.add(i);
         List<Integer> first = buf.drain(idx);
@@ -46,7 +46,7 @@ public class CircularBufferTest {
     @Test
     public void whatHappensAtWraparound() throws Exception {
         CircularBuffer<Integer> buf = new CircularBuffer<Integer>(10);
-        AtomicInteger idx = buf.index();
+        AtomicLong idx = buf.index();
         buf.add(1);
         assumeThat(buf.take(idx), is(1));
         for (int i = 2; i < buf.size()+3; i++)
